Block (unknown) being interpreted as a HEAD branch name
In a merge commit being created in a pull request
(https://github.com/stan-dev/stan-mode/pull/58),
the HEAD branch was "(unknown)". This was taken literally
as a branch name.

This solution to another problem essentially assumes
the master branch when the HEAD branch is ambiguous.
https://github.com/melpa/melpa/issues/300

This commit turns "(unknown)" into nil, so that the default
"master" is used.